refactor(pipe): Split pipeline into core and specialized modules
This refactor separates the shared pipeline engine from specific
segmenter implementations. Shared logic for step initialization,
state management, and execution is moved to `pipe_core.py`, while
Cellpose and BABY-specific logic are isolated in their own modules.

- Extract shared engine and helper functions to `pipe_core.py`
- Create `pipe_baby.py` and `pipe_builder_baby.py` for BABY
- pipelines Simplify `pipe.py` and `pipe_builder.py` to focus on
- Cellpose Support pipeline-specific post-state hooks for custom
- metadata extraction
